// src/sequelize/types.ts
export interface KishiModelAttributeColumnOptions extends ModelAttributeColumnOptions<KishiModel> {
  /**
   * can be updated
   */
  update?: boolean;
  /**
   * model names of custom dependencies
   * TODO
   */
  dependencies?: string[];
  /**
   * custom function for view, or false for invisible attributes
   */
  toView?: boolean | ((value: unknown) => unknown);
  /**
   * custom function from view, or false for not assignable from api
   */
  fromView?: boolean | ((value: unknown) => unknown);
  /**
   * custom where function
   * @example 
   * where:{"jsonField":"value"}=>where:{"#substring":"\"jsonField\":\"value\""}
   * //TODO
   */
  where?: boolean | ((value: WhereOptions<any>) => WhereOptions<any>);
  /**
   * the literal field type, used for auto generating typescript interfaces
   * @example "string","Number","any"
   */
  ts_typeStr?: string | ((imports: [string, string][]) => string);
  /**
   * bind this attribute to an association field
   * @example "string","Number","any"
   */
  binder?: TorArray<{
    associationName: string;
    targetField: string;
    sourceField?: string;
    hardBind?: boolean;
  }>
  counter?: {
    associationName: string;
    targetField: string;
    sourceField?: string;
  }
}